backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
motor==3.3.2
pydantic==2.5.3
python-dotenv==1.0.0
google-generativeai>=0.7.0,<0.8.0
pymongo==4.6.1
pytest==7.4.3
httpx==0.26.0
pytz==2024.1
langgraph==0.2.28
langchain-core==0.3.6
langchain-google-genai==2.0.0
